CS-27: introduce be_not_success matcher
