+++ lib/convenient_service/rspec/matchers/custom/results.rb
@@ -1,5 +1,7 @@
 # frozen_string_literal: true
 
+require_relative "results/base"
 require_relative "results/be_error"
 require_relative "results/be_failure"
+require_relative "results/be_not_success"
 require_relative "results/be_success"
