+++ lib/convenient_service/rspec/matchers/custom/results/be_not_success.rb
@@ -5,9 +5,12 @@ module ConvenientService
     module Matchers
       module Custom
         module Results
-          class BeFailure < Results::Base
+          class BeNotSuccess < Results::Base
             def statuses
-              [Service::Plugins::HasResult::Constants::FAILURE_STATUS]
+              [
+                Service::Plugins::HasResult::Constants::ERROR_STATUS,
+                Service::Plugins::HasResult::Constants::FAILURE_STATUS
+              ]
             end
           end
         end
