+++ lib/convenient_service/service/plugins/has_result/constants.rb
@@ -5,10 +5,7 @@ module ConvenientService
     module Plugins
       module HasResult
         module Constants
-          ##
-          # TODO: Consider to include `::ConvenientService::Configs::Minimal`
-          #
-          DEFAULT_SERVICE_CLASS = ::Class.new { include ::ConvenientService::Core }
+          DEFAULT_SERVICE_CLASS = ::Class.new
           DEFAULT_SERVICE_INSTANCE = DEFAULT_SERVICE_CLASS.new
 
           SUCCESS_STATUS = :success
