+++ spec/lib/convenient_service/rspec/matchers/custom/results/base_spec.rb
@@ -1,0 +1,3 @@
+##
+# TODO: Specs.
+#
